requirements.txt
langchain
langchain-groq
langchain-community
python-dotenv
langchain-huggingface
langchain-text-splitters
langchain-core